+++ src/common/NetworkManager.ts
@@ -22,11 +22,7 @@ import { helper, debugError } from './helper.js';
 import { Protocol } from 'devtools-protocol';
 import { HTTPRequest } from './HTTPRequest.js';
 import { HTTPResponse } from './HTTPResponse.js';
-import {
-  FetchRequestId,
-  NetworkEventManager,
-  NetworkRequestId,
-} from './NetworkEventManager.js';
+import { FetchRequestId, NetworkEventManager } from './NetworkEventManager.js';
 
 /**
  * @public
@@ -172,9 +168,7 @@ export class NetworkManager extends EventEmitter {
   }
 
   numRequestsInProgress(): number {
-    return [...this._networkEventManager.httpRequest].filter(([, request]) => {
-      return !request.response();
-    }).length;
+    return this._networkEventManager.numRequestsInProgress();
   }
 
   async setOfflineMode(value: boolean): Promise<void> {
@@ -265,17 +259,17 @@ export class NetworkManager extends EventEmitter {
     ) {
       const { requestId: networkRequestId } = event;
 
-      this._networkEventManager.requestWillBeSent.set(networkRequestId, event);
+      this._networkEventManager.storeRequestWillBeSent(networkRequestId, event);
 
       /**
        * CDP may have sent a Fetch.requestPaused event already. Check for it.
        */
       const requestPausedEvent =
-        this._networkEventManager.requestPaused.get(networkRequestId);
+        this._networkEventManager.getRequestPaused(networkRequestId);
       if (requestPausedEvent) {
         const { requestId: fetchRequestId } = requestPausedEvent;
         this._onRequest(event, fetchRequestId);
-        this._networkEventManager.requestPaused.delete(networkRequestId);
+        this._networkEventManager.forgetRequestPaused(networkRequestId);
       }
 
       return;
@@ -336,7 +330,7 @@ export class NetworkManager extends EventEmitter {
 
     const requestWillBeSentEvent = (() => {
       const requestWillBeSentEvent =
-        this._networkEventManager.requestWillBeSent.get(networkRequestId);
+        this._networkEventManager.getRequestWillBeSent(networkRequestId);
 
       // redirect requests have the same `requestId`,
       if (
@@ -344,7 +338,7 @@ export class NetworkManager extends EventEmitter {
         (requestWillBeSentEvent.request.url !== event.request.url ||
           requestWillBeSentEvent.request.method !== event.request.method)
       ) {
-        this._networkEventManager.requestWillBeSent.delete(networkRequestId);
+        this._networkEventManager.forgetRequestWillBeSent(networkRequestId);
         return;
       }
       return requestWillBeSentEvent;
@@ -353,7 +347,7 @@ export class NetworkManager extends EventEmitter {
     if (requestWillBeSentEvent) {
       this._onRequest(requestWillBeSentEvent, fetchRequestId);
     } else {
-      this._networkEventManager.requestPaused.set(networkRequestId, event);
+      this._networkEventManager.storeRequestPaused(networkRequestId, event);
     }
   }
 
@@ -376,7 +370,7 @@ export class NetworkManager extends EventEmitter {
           .responseExtraInfo(event.requestId)
           .shift();
         if (!redirectResponseExtraInfo) {
-          this._networkEventManager.queuedRedirectInfo(event.requestId).push({
+          this._networkEventManager.queueRedirectInfo(event.requestId, {
             event,
             fetchRequestId,
           });
@@ -384,9 +378,7 @@ export class NetworkManager extends EventEmitter {
         }
       }
 
-      const request = this._networkEventManager.httpRequest.get(
-        event.requestId
-      );
+      const request = this._networkEventManager.getRequest(event.requestId);
       // If we connect late to the target, we could have missed the
       // requestWillBeSent event.
       if (request) {
@@ -409,7 +401,7 @@ export class NetworkManager extends EventEmitter {
       event,
       redirectChain
     );
-    this._networkEventManager.httpRequest.set(event.requestId, request);
+    this._networkEventManager.storeRequest(event.requestId, request);
     this.emit(NetworkManagerEmittedEvents.Request, request);
     request.finalizeInterceptions();
   }
@@ -417,7 +409,7 @@ export class NetworkManager extends EventEmitter {
   _onRequestServedFromCache(
     event: Protocol.Network.RequestServedFromCacheEvent
   ): void {
-    const request = this._networkEventManager.httpRequest.get(event.requestId);
+    const request = this._networkEventManager.getRequest(event.requestId);
     if (request) request._fromMemoryCache = true;
     this.emit(NetworkManagerEmittedEvents.RequestServedFromCache, request);
   }
@@ -447,7 +439,7 @@ export class NetworkManager extends EventEmitter {
     responseReceived: Protocol.Network.ResponseReceivedEvent,
     extraInfo: Protocol.Network.ResponseReceivedExtraInfoEvent | null
   ): void {
-    const request = this._networkEventManager.httpRequest.get(
+    const request = this._networkEventManager.getRequest(
       responseReceived.requestId
     );
     // FileUpload sends a response without a matching request.
@@ -476,7 +468,7 @@ export class NetworkManager extends EventEmitter {
   }
 
   _onResponseReceived(event: Protocol.Network.ResponseReceivedEvent): void {
-    const request = this._networkEventManager.httpRequest.get(event.requestId);
+    const request = this._networkEventManager.getRequest(event.requestId);
     let extraInfo = null;
     if (request && !request._fromMemoryCache && event.hasExtraInfo) {
       extraInfo = this._networkEventManager
@@ -484,26 +476,13 @@ export class NetworkManager extends EventEmitter {
         .shift();
       if (!extraInfo) {
         // Wait until we get the corresponding ExtraInfo event.
-        let resolver = null;
-        const promise = new Promise<void>((resolve) => (resolver = resolve));
-        this._networkEventManager.queuedEvents.set(event.requestId, {
-          responseReceived: event,
-          promise,
-          resolver,
+        this._networkEventManager.queueEventGroup(event.requestId, {
+          responseReceivedEvent: event,
         });
         return;
       }
     }
     this._emitResponseEvent(event, extraInfo);
-  }
-
-  responseWaitingForExtraInfoPromise(
-    networkRequestId: NetworkRequestId
-  ): Promise<void> {
-    const responseReceived =
-      this._networkEventManager.queuedEvents.get(networkRequestId);
-    if (!responseReceived) return Promise.resolve();
-    return responseReceived.promise;
   }
 
   _onResponseReceivedExtraInfo(
@@ -512,9 +491,9 @@ export class NetworkManager extends EventEmitter {
     // We may have skipped a redirect response/request pair due to waiting for
     // this ExtraInfo event. If so, continue that work now that we have the
     // request.
-    const redirectInfo = this._networkEventManager
-      .queuedRedirectInfo(event.requestId)
-      .shift();
+    const redirectInfo = this._networkEventManager.takeQueuedRedirectInfo(
+      event.requestId
+    );
     if (redirectInfo) {
       this._networkEventManager.responseExtraInfo(event.requestId).push(event);
       this._onRequest(redirectInfo.event, redirectInfo.fetchRequestId);
@@ -523,18 +502,17 @@ export class NetworkManager extends EventEmitter {
 
     // We may have skipped response and loading events because we didn't have
     // this ExtraInfo event yet. If so, emit those events now.
-    const queuedEvents = this._networkEventManager.queuedEvents.get(
+    const queuedEvents = this._networkEventManager.getQueuedEventGroup(
       event.requestId
     );
     if (queuedEvents) {
-      this._emitResponseEvent(queuedEvents.responseReceived, event);
-      if (queuedEvents.loadingFinished) {
-        this._emitLoadingFinished(queuedEvents.loadingFinished);
+      this._emitResponseEvent(queuedEvents.responseReceivedEvent, event);
+      if (queuedEvents.loadingFinishedEvent) {
+        this._emitLoadingFinished(queuedEvents.loadingFinishedEvent);
       }
-      if (queuedEvents.loadingFailed) {
-        this._emitLoadingFailed(queuedEvents.loadingFailed);
+      if (queuedEvents.loadingFailedEvent) {
+        this._emitLoadingFailed(queuedEvents.loadingFailedEvent);
       }
-      queuedEvents.resolver();
       return;
     }
 
@@ -546,7 +524,7 @@ export class NetworkManager extends EventEmitter {
     const requestId = request._requestId;
     const interceptionId = request._interceptionId;
 
-    this._networkEventManager.httpRequest.delete(requestId);
+    this._networkEventManager.forgetRequest(requestId);
     this._attemptedAuthentications.delete(interceptionId);
 
     if (events) {
@@ -557,18 +535,18 @@ export class NetworkManager extends EventEmitter {
   _onLoadingFinished(event: Protocol.Network.LoadingFinishedEvent): void {
     // If the response event for this request is still waiting on a
     // corresponding ExtraInfo event, then wait to emit this event too.
-    const queuedEvents = this._networkEventManager.queuedEvents.get(
+    const queuedEvents = this._networkEventManager.getQueuedEventGroup(
       event.requestId
     );
     if (queuedEvents) {
-      queuedEvents.loadingFinished = event;
+      queuedEvents.loadingFinishedEvent = event;
     } else {
       this._emitLoadingFinished(event);
     }
   }
 
   _emitLoadingFinished(event: Protocol.Network.LoadingFinishedEvent): void {
-    const request = this._networkEventManager.httpRequest.get(event.requestId);
+    const request = this._networkEventManager.getRequest(event.requestId);
     // For certain requestIds we never receive requestWillBeSent event.
     // @see https://crbug.com/750469
     if (!request) return;
@@ -583,18 +561,18 @@ export class NetworkManager extends EventEmitter {
   _onLoadingFailed(event: Protocol.Network.LoadingFailedEvent): void {
     // If the response event for this request is still waiting on a
     // corresponding ExtraInfo event, then wait to emit this event too.
-    const queuedEvents = this._networkEventManager.queuedEvents.get(
+    const queuedEvents = this._networkEventManager.getQueuedEventGroup(
       event.requestId
     );
     if (queuedEvents) {
-      queuedEvents.loadingFailed = event;
+      queuedEvents.loadingFailedEvent = event;
     } else {
       this._emitLoadingFailed(event);
     }
   }
 
   _emitLoadingFailed(event: Protocol.Network.LoadingFailedEvent): void {
-    const request = this._networkEventManager.httpRequest.get(event.requestId);
+    const request = this._networkEventManager.getRequest(event.requestId);
     // For certain requestIds we never receive requestWillBeSent event.
     // @see https://crbug.com/750469
     if (!request) return;