+++ test/elementhandle.spec.ts
@@ -257,6 +257,29 @@ describe('ElementHandle specs', function () {
     });
   });
 
+  describe('Element.waitForSelector', () => {
+    it('should wait correctly with waitForSelector on an element', async () => {
+      const { page } = getTestState();
+      const waitFor = page.waitForSelector('.foo');
+      // Set the page content after the waitFor has been started.
+      await page.setContent(
+        '<div id="not-foo"></div><div class="bar">bar2</div><div class="foo">Foo1</div>'
+      );
+      let element = await waitFor;
+      expect(element).toBeDefined();
+
+      const innerWaitFor = element.waitForSelector('.bar');
+      await element.evaluate((el) => {
+        el.innerHTML = '<div class="bar">bar1</div>';
+      });
+      element = await innerWaitFor;
+      expect(element).toBeDefined();
+      expect(
+        await element.evaluate((el: HTMLElement) => el.innerText)
+      ).toStrictEqual('bar1');
+    });
+  });
+
   describe('ElementHandle.hover', function () {
     it('should work', async () => {
       const { page, server } = getTestState();
@@ -417,6 +440,33 @@ describe('ElementHandle specs', function () {
       const element = await waitFor;
 
       expect(element).toBeDefined();
+    });
+
+    it('should wait correctly with waitForSelector on an element', async () => {
+      const { page, puppeteer } = getTestState();
+      puppeteer.registerCustomQueryHandler('getByClass', {
+        queryOne: (element, selector) => element.querySelector(`.${selector}`),
+      });
+      const waitFor = page.waitForSelector('getByClass/foo');
+
+      // Set the page content after the waitFor has been started.
+      await page.setContent(
+        '<div id="not-foo"></div><div class="bar">bar2</div><div class="foo">Foo1</div>'
+      );
+      let element = await waitFor;
+      expect(element).toBeDefined();
+
+      const innerWaitFor = element.waitForSelector('getByClass/bar');
+
+      await element.evaluate((el) => {
+        el.innerHTML = '<div class="bar">bar1</div>';
+      });
+
+      element = await innerWaitFor;
+      expect(element).toBeDefined();
+      expect(
+        await element.evaluate((el: HTMLElement) => el.innerText)
+      ).toStrictEqual('bar1');
     });
 
     it('should wait correctly with waitFor', async () => {