+++ test/requestinterception-experimental.spec.ts
@@ -25,6 +25,7 @@ import {
   describeFailsFirefox,
 } from './mocha-utils'; // eslint-disable-line import/extensions
 import { ActionResult } from '../lib/cjs/puppeteer/api-docs-entry.js';
+import { InterceptResolutionAction } from '../lib/cjs/puppeteer/common/HTTPRequest.js';
 
 describe('request interception', function () {
   setupTestBrowserHooks();
@@ -845,7 +846,7 @@ describe('request interception', function () {
         'Yo, page!'
       );
     });
-    it('should indicate alreay-handled if an intercept has been handled', async () => {
+    it('should indicate already-handled if an intercept has been handled', async () => {
       const { page, server } = getTestState();
 
       await page.setRequestInterception(true);
@@ -857,7 +858,7 @@ describe('request interception', function () {
       });
       page.on('request', (request) => {
         const { action } = request.interceptResolutionState();
-        expect(action).toBe('alreay-handled');
+        expect(action).toBe(InterceptResolutionAction.AlreadyHandled);
       });
       await page.goto(server.EMPTY_PAGE);
     });